// generated/models/ApiDropCompetitionContext.ts
// @ts-nocheck
/**
 * 6529.io API
 * JSON REST API for 6529.io. New here? A terminology overview and a step-by-step authentication walkthrough (guides &amp; auth) live at <a href=\"https://6529.io/tools/api\">https://6529.io/tools/api</a>. The raw machine-readable spec is downloadable at <a href=\"/openapi.yaml\">/openapi.yaml</a> and <a href=\"/openapi.json\">/openapi.json</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ApiCompetition } from '../models/ApiCompetition';
import { ApiCompetitionDropVoteSummary } from '../models/ApiCompetitionDropVoteSummary';
import { ApiCompetitionEntry } from '../models/ApiCompetitionEntry';
import { HttpFile } from '../http/http';

export class ApiDropCompetitionContext {
    'vote_summary'?: ApiCompetitionDropVoteSummary | null;
    'competition': ApiCompetition | null;
    'entry': ApiCompetitionEntry | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "vote_summary",
            "baseName": "vote_summary",
            "type": "ApiCompetitionDropVoteSummary",
            "format": ""
        },
        {
            "name": "competition",
            "baseName": "competition",
            "type": "ApiCompetition",
            "format": ""
        },
        {
            "name": "entry",
            "baseName": "entry",
            "type": "ApiCompetitionEntry",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiDropCompetitionContext.attributeTypeMap;
    }

    public constructor() {
    }
}