// generated/models/ApiCompetitionDistributionItemPage.ts
// @ts-nocheck
/**
 * 6529.io API
 * JSON REST API for 6529.io. New here? A terminology overview and a step-by-step authentication walkthrough (guides &amp; auth) live at <a href=\"https://6529.io/tools/api\">https://6529.io/tools/api</a>. The raw machine-readable spec is downloadable at <a href=\"/openapi.yaml\">/openapi.yaml</a> and <a href=\"/openapi.json\">/openapi.json</a>.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ApiCompetitionDistributionItem } from '../models/ApiCompetitionDistributionItem';
import { HttpFile } from '../http/http';

export class ApiCompetitionDistributionItemPage {
    'data': Array<ApiCompetitionDistributionItem>;
    'next_cursor': string | null;
    'has_more': boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "data",
            "baseName": "data",
            "type": "Array<ApiCompetitionDistributionItem>",
            "format": ""
        },
        {
            "name": "next_cursor",
            "baseName": "next_cursor",
            "type": "string",
            "format": ""
        },
        {
            "name": "has_more",
            "baseName": "has_more",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ApiCompetitionDistributionItemPage.attributeTypeMap;
    }

    public constructor() {
    }
}